pyproject.toml
[tool.poetry]
name = "motosell"
version = "0.1.0"
description = ""
authors = ["Sebastian <sebastian.raczylo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1"
django-rest-framework = "^0.1.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
django-crispy-forms = "^1.14.0"
Pillow = "^9.2.0"
djangorestframework-simplejwt = "^5.2.0"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
